+++ src/Common/ninjam/client/ServerMessages.cpp
@@ -446,7 +446,7 @@ ServerToClientChatMessage ServerToClientChatMessage::from(QIODevice *device, qui
 
     int readed = stream.readRawData(byteArray.data(), payload);
 
-    Q_ASSERT(readed == payload);
+    Q_ASSERT(readed == (int)payload);
 
     auto arrays = byteArray.split('\0');
     Q_ASSERT(arrays.size() >= 5);
